feedback_service/requirements.txt
fastapi
uvicorn
python-dotenv
asyncmy